refactor(v2): scope getFinalRoutes to docusaurus package
