+++ packages/docusaurus/src/server/__tests__/utils.test.ts
@@ -1,0 +1,29 @@
+/**
+ * Copyright (c) Facebook, Inc. and its affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+import flatMap from 'lodash.flatmap';
+import {RouteConfig} from '@docusaurus/types';
+import {getFinalRoutes} from '../utils';
+
+describe('getFinalRoutes', () => {
+  test('should get final routes correctly', () => {
+    const routes: RouteConfig[] = [
+      {
+        path: '/docs',
+        component: '',
+        routes: [
+          {path: '/docs/someDoc', component: ''},
+          {path: '/docs/someOtherDoc', component: ''},
+        ],
+      },
+      {
+        path: '/community',
+        component: '',
+      },
+    ];
+    expect(flatMap(routes, getFinalRoutes)).toMatchSnapshot();
+  });
+});
