+++ packages/docusaurus/src/server/plugins/init.ts
@@ -23,7 +23,7 @@ import * as Joi from '@hapi/joi';
 import {
   isValidationDisabledEscapeHatch,
   logValidationBugReportHint,
-} from '../validationUtils';
+} from '../utils';
 
 function pluginOptionsValidator<T>(
   schema: ValidationSchema<T>,
